Add Galaga port, host test harnesses, and input debouncing
Galaga (ArcadeMachine_Galaga + galaga_fruitjam) is the project's first
multi-CPU machine: three Z80s sharing RAM, a Namco 06XX/51XX/54XX custom
I/O chain, and a fourth video layer (the 05XX starfield). Confirmed on
real hardware at a flat 60fps with ~3ms of frame-budget headroom --
starfield, sprites, joystick/fire, WSG music and 54XX explosions.

Hardware facts are cited against MAME's galaga driver, its device sources
and its discrete audio netlist. The 51XX and 54XX are hand-written HLE
because modern MAME emulates them at low level, running MB8843/MB8844
firmware this project has no dump of; both files say plainly which parts
are cited and which are approximated.

Also adds:

- tools/ -- host test harnesses that build ArcadeMachine_* libraries into
  native executables running the real machine against a stub ArcadeHAL.
  This works only because SAMP's architecture rule holds. A hardware
  iteration costs minutes; this costs about a second, with frame
  rendering to PPM, audio capture to WAV, and sprite census.
- Input debouncing in hal_input_fruitjam.cpp. The project previously had
  none anywhere; Bounce2 was listed as a dependency but unused (now
  dropped from the install instructions -- verified by building without
  it). Asymmetric: press believed instantly, release must persist 25ms.
- input_bounce_test_fruitjam -- 10kHz contact profiler that replays each
  press through a simulated 60Hz sampler and counts press edges.
- Pac-Man and Galaga now default to rotation 3, not 1. The 8080bw and
  Namco cabinets mounted their monitors opposite ways, so the two
  families need opposite software rotations to be upright on one screen.
  Both Namco machines had copied Invaders' default. See DEVNOTES #33.

DEVNOTES gains problems #24-34, the bring-up record for all of the above.
Several are lessons rather than bugs -- #32 in particular documents a
flawed harness experiment that compared several equally-wrong runs to
each other, agreed with itself, and cost a session.

KNOWN ISSUE, diagnosed but NOT fixed in this commit: Lunar Rescue still
shows red lines during peak simultaneous audio (most reliably the bonus
arpeggio when an astronaut is returned). See DEVNOTES #34. The cause is
that lrescue_run_frame() runs the whole frame's i8080 emulation in one
uninterrupted burst before submitting any scanline -- the same bug #20
fixed for Pac-Man, never back-applied to the i8080 games. Core 1 coasts
on the 8-buffer queue plus vblank, about 2ms, leaving ~200us of margin,
which a 232us audio ISR exceeds.

An interim mitigation is included: BUFFER_SAMPLES halved 256->128, which
halves the ISR's blocking window at no cost in total CPU. Measurably
better on hardware, not eliminated. The real fix -- interleaving CPU
execution with the scanline pump -- is next; DEVNOTES #34 records the
plan and the three things to watch (mid-frame render state, interrupt
cycle placement, and the cyc carry that keeps the audio clock aligned).
Invaders has the identical structure and should get the same treatment.

+++ libraries/ArcadeCPU_Z80/src/z80.h
@@ -33,7 +33,16 @@ struct z80 {
   void (*port_out)(z80*, uint8_t, uint8_t);
   void* userdata;
 
-  unsigned long cyc; // cycle count (t-states)
+  // Divergence from upstream superzazu/z80 (the third; see z80.c's
+  // Z80_RAMFUNC and this file's extern "C" guard): pinned to an explicit
+  // 32-bit type rather than `unsigned long`. `unsigned long` is 32-bit on
+  // the arm-none-eabi target but 64-bit on a typical host, which made the
+  // host test harness (arcade_arduino/tools/galaga_host) structurally
+  // unable to reproduce cycle-counter wraparound -- and a real freeze bug
+  // hid in exactly that gap for a while (see galaga_machine.cpp's
+  // interleave_to_target()). Machines deriving timing from this counter
+  // should use uint32_t to match.
+  uint32_t cyc; // cycle count (t-states)
 
   uint16_t pc, sp, ix, iy; // special purpose registers
   uint16_t mem_ptr; // "wz" register
